+++ docs/demo/select-tree/section-multiple.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import SelectTree from '../../../components/select-tree'
 const prefix = 'tree-select-multiple'
-const rightOptions = ['基础', '默认值', '数据回显', '默认展开']
+const rightOptions = ['基础', '默认值', '数据回显', '默认展开', '数据转换']
 const desc = '展示从多个收起的备选项中选出的一个选项'
 const defaultStr = `constructor () {
   super()
@@ -140,8 +140,74 @@ class Demo extends React.Component {
     code: `import React from 'react'
     import SelectTree from '@hi-ui/hiui/es/select-tree'\n
     class Demo extends React.Component {
-      ${defaultStr}
-
+      constructor () {
+        super()
+        this.state = {
+          value: ['0-0'],
+          singleList: [
+            {
+              content: '手机类',
+              id: '0',
+              children: [
+                {
+                  content: 'Redmi系列',
+                  id: '0-0',
+                  disabled: true,
+                  children: [{
+                    id: '0-0-1',
+                    content: 'Redmi K30'
+                  }, {
+                    id: '0-0-2',
+                    content: 'Redmi K30 Pro'
+                  }, {
+                    id: '0-0-3',
+                    content: 'Redmi 10X 5G'
+                  }, {
+                    id: '0-0-4',
+                    content: 'Redmi Note 8'
+                  }, {
+                    id: '0-0-5',
+                    content: 'Redmi 9'
+                  }, {
+                    id: '0-0-6',
+                    content: 'Redmi 9A'
+                  }]
+                },{
+                  content: '小米手机',
+                  id: '0-1',
+                  children: [{
+                    id: '0-1-1',
+                    content: '小米10 Pro'
+                  }, {
+                    id: '0-1-2',
+                    content: '小米10'
+                  }, {
+                    id: '0-1-3',
+                    content: '小米10 青春版 5G'
+                  }, {
+                    id: '0-1-4',
+                    content: '小米MIX Alpha'
+                  }]
+                },
+              ]
+            },
+            {
+              content: '电视',
+              id: '1',
+              children: [{
+                content: '小米电视 大师 65英寸OLED',
+                id: '1-0'
+              }, {
+                content: 'Redmi 智能电视 MAX 98',
+                id: '1-1'
+              }, {
+                content: '小米电视4A 60英寸',
+                id: '1-2'
+              }]
+            }
+          ]
+        }
+      }
       render () {
         const { singleList } = this.state
         return (
@@ -149,12 +215,13 @@ class Demo extends React.Component {
             clearable
             type='multiple'
             data={singleList}
+            fieldNames={{title: 'content'}}
             defaultExpandIds={['1-0']}
           />
         )
       }
     }`,
-    opt: ['默认展开']
+    opt: ['数据转换']
   }
 ]
 const DemoType = () => (